Update jupyter-widget imports for JupyterLab support
