+++ modules/jupyter-widget/src/notebook-widget.js
@@ -1,0 +1,6 @@
+import {DeckGLView, DeckGLModel} from './widget';
+import {initDeck} from './notebook-utils';
+
+DeckGLView.deckInitFunction = initDeck;
+
+export {DeckGLView, DeckGLModel};
