+++ modules/jupyter-widget/src/plugin.js
@@ -3,11 +3,8 @@
  * See:
  * https://github.com/jupyter-widgets/widget-ts-cookiecutter/blob/51e9fed8687e3b9cf1ed2fd307b7675e864f89ae/%7B%7Bcookiecutter.github_project_name%7D%7D/src/plugin.ts
  */
-
 import {IJupyterWidgetRegistry} from '@jupyter-widgets/base';
-
-import * as widgetExports from './widget';
-
+import {DeckGLView, DeckGLModel} from './lab-widget';
 import {MODULE_NAME, MODULE_VERSION} from './version';
 
 const EXTENSION_ID = '@deck.gl/jupyter-widget:plugin';
@@ -25,6 +22,6 @@ function activateWidgetExtension(app, registry) {
   registry.registerWidget({
     name: MODULE_NAME,
     version: MODULE_VERSION,
-    exports: widgetExports
+    exports: {DeckGLView, DeckGLModel}
   });
 }
